+++ src/global.hpp
@@ -13,10 +13,10 @@
 
 extern body *pc;
 extern body *w;
-extern int32_t *stack;
+extern intptr_t *stack;
 extern body **rstack;
 
-extern int32_t stack_start[FORTH_STACK_SIZE];
+extern intptr_t stack_start[FORTH_STACK_SIZE];
 extern body *rstack_start[FORTH_RSTACK_SIZE];
 extern char input_buf[FORTH_INPUT_BUF];
 extern int32_t user_mem[FORTH_USER_MEM];
