+++ src/words.cpp
@@ -8,7 +8,7 @@ static body program_stub = {.inner = w_main_stub.xt};
  *
  * @param value 値
  */
-static void push(int32_t value)
+static void push(intptr_t value)
 {
     stack++;
     *stack = value;
@@ -18,11 +18,11 @@ static void push(int32_t value)
 /**
  * @brief スタックから値をポップする。
  *
- * @return int32_t 値
+ * @return intptr_t 値
  */
-static int32_t pop()
+static intptr_t pop()
 {
-    int32_t value = *stack;
+    intptr_t value = *stack;
     stack--;
     return value;
 }
@@ -118,7 +118,7 @@ native1(drop)
  */
 native1(swap)
 {
-    int32_t a, b;
+    intptr_t a, b;
     a = pop();
     b = pop();
     push(a);
@@ -134,7 +134,7 @@ native1(swap)
  */
 native1(dup)
 {
-    int32_t a;
+    intptr_t a;
     a = pop();
     push(a);
     push(a);
@@ -149,7 +149,7 @@ native1(dup)
  */
 native1(rot)
 {
-    int32_t x1, x2, x3;
+    intptr_t x1, x2, x3;
     x1 = pop();
     x2 = pop();
     x3 = pop();
@@ -165,7 +165,7 @@ native1(rot)
  */
 native2("+", plus)
 {
-    int32_t x1, x2;
+    intptr_t x1, x2;
     x1 = pop();
     x2 = pop();
     push(x1 + x2);
@@ -178,7 +178,7 @@ native2("+", plus)
  */
 native2("-", minus)
 {
-    int32_t x1, x2;
+    intptr_t x1, x2;
     x1 = pop();
     x2 = pop();
     push(x1 - x2);
@@ -191,7 +191,7 @@ native2("-", minus)
  */
 native2("*", mul)
 {
-    int32_t x1, x2;
+    intptr_t x1, x2;
     x1 = pop();
     x2 = pop();
     push(x1 * x2);
@@ -204,7 +204,7 @@ native2("*", mul)
  */
 native2("/", div)
 {
-    int32_t x1, x2;
+    intptr_t x1, x2;
     x1 = pop();
     x2 = pop();
     push(x1 / x2);
@@ -217,7 +217,7 @@ native2("/", div)
  */
 native1(not )
 {
-    int32_t x;
+    intptr_t x;
     x = pop();
     push(!x);
     next();
@@ -229,7 +229,7 @@ native1(not )
  */
 native2("=", eq)
 {
-    int32_t x1, x2;
+    intptr_t x1, x2;
     x1 = pop();
     x2 = pop();
     push(x1 == x2);
@@ -241,7 +241,7 @@ native2("=", eq)
  */
 native2(".s", show_stack)
 {
-    for (int32_t *i = stack; i > stack_start; i--)
+    for (intptr_t *i = stack; i > stack_start; i--)
     {
 #if defined(LOCAL)
         printf("%d\n", *i);
